refactor(tests): isolate fixture upgrade contract
